Refactor Python backend into MVC structure
