+++ python-backend/app/__init__.py
@@ -1,0 +1,5 @@
+"""Expose the FastAPI application instance for ASGI servers."""
+
+from .main import app
+
+__all__ = ["app"]
