+++ python-backend/app/models/__init__.py
@@ -1,0 +1,7 @@
+"""Database models for the FastAPI application."""
+
+from .embedding import Embedding
+from .image import ImageMetadata
+from .user import User
+
+__all__ = ["Embedding", "ImageMetadata", "User"]
