+++ python-backend/app/models/image.py
@@ -1,25 +1,14 @@
+"""Image metadata model definition."""
+
 from sqlalchemy import Column, Integer, String
 from pgvector.sqlalchemy import Vector
-from .db import Base
 
-
-class User(Base):
-    __tablename__ = "users"
-
-    id = Column(Integer, primary_key=True, index=True)
-    username = Column(String, unique=True, index=True)
-    hashed_password = Column(String)
-
-
-class Embedding(Base):
-    __tablename__ = "embeddings"
-
-    id = Column(Integer, primary_key=True, index=True)
-    content = Column(String, nullable=True)
-    embedding = Column(Vector(3))
+from ..db import Base
 
 
 class ImageMetadata(Base):
+    """Stores metadata and embeddings for uploaded images."""
+
     __tablename__ = "images"
 
     id = Column(Integer, primary_key=True, index=True)
