+++ python-backend/app/views/__init__.py
@@ -1,0 +1,30 @@
+"""Public exports for API schemas and helpers."""
+
+from .schemas import (
+    AskRequest,
+    AskResponse,
+    EmbeddingOut,
+    ImageOut,
+    OpenAIRequest,
+    Token,
+    TokenData,
+    UserBase,
+    UserCreate,
+    UserOut,
+)
+from .transformers import embedding_to_out, image_to_out
+
+__all__ = [
+    "AskRequest",
+    "AskResponse",
+    "EmbeddingOut",
+    "ImageOut",
+    "OpenAIRequest",
+    "Token",
+    "TokenData",
+    "UserBase",
+    "UserCreate",
+    "UserOut",
+    "embedding_to_out",
+    "image_to_out",
+]
